Moved some view methods to the activity
Show and Hide are actions that the presenter should tell the activity to do, and not do it himself
